+++ resources/js/Pages/Search/Index.vue
@@ -3,28 +3,21 @@ import Card from '@/Components/Card.vue';
 import FormInput from '@/Components/Forms/FormInput.vue';
 import FormCheckbox from '@/Components/Forms/FormCheckbox.vue';
 import useSearch from '@/composables/useSearch';
-import {
-  EaterySearchResult,
-  SearchableItem,
-  SearchParams,
-  SearchResult,
-} from '@/types/Search';
+import { SearchParams, SearchResult } from '@/types/Search';
 import { PaginatedResponse } from '@/types/GenericTypes';
 import { nextTick, onMounted, Ref, ref, watch } from 'vue';
 import { watchDebounced } from '@vueuse/core';
 import Loader from '@/Components/Loader.vue';
-import RecipeSquareImage from '@/Components/PageSpecific/Recipes/RecipeSquareImage.vue';
-import { Link, router } from '@inertiajs/vue3';
-import StaticMap from '@/Components/Maps/StaticMap.vue';
-import { pluralise } from '@/helpers';
+import { Deferred, router } from '@inertiajs/vue3';
 import eventBus from '@/eventBus';
-import useInfiniteScrollCollection from '@/composables/useInfiniteScrollCollection';
 import useBrowser from '@/composables/useBrowser';
+import SearchResults from '@/Components/PageSpecific/SearchResults.vue';
+import { VisitOptions } from '@inertiajs/core/types/types';
 
 const props = defineProps<{
   parameters: SearchParams;
   location: string;
-  results: PaginatedResponse<SearchResult>;
+  results?: PaginatedResponse<SearchResult>;
   hasEatery: boolean;
   aiAssisted: boolean;
 }>();
@@ -41,26 +34,42 @@ const formParamsToSearchParams = (): URLSearchParams => {
 
 const landmark: Ref<Element> = ref();
 
+const resultsElem: Ref<null | {
+  reset: () => void;
+  pause: boolean;
+  refreshUrl: (url: string) => void;
+  requestOptions: Partial<VisitOptions>;
+}> = ref(null);
+
 const { hasError, searchForm, latLng, submitSearch } = useSearch();
 
-const { items, pause, reset, refreshUrl, requestOptions } =
-  useInfiniteScrollCollection<SearchResult>('results', landmark);
-
 onMounted(() => {
-  pause.value = true;
+  if (resultsElem.value) {
+    resultsElem.value.pause = true;
+  }
 
   if (props.aiAssisted) {
     const url = new URL(useBrowser().currentUrl());
     url.search = formParamsToSearchParams().toString();
 
     nextTick(() => {
-      history.pushState(null, '', url.toString());
-      refreshUrl(url.pathname + url.search);
+      router.get(
+        url.toString(),
+        {},
+        {
+          preserveScroll: true,
+          preserveState: true,
+          replace: true,
+          onSuccess: () => {
+            resultsElem.value?.refreshUrl(url.pathname + url.search);
+          },
+        },
+      );
     });
   }
 
-  if (latLng.value) {
-    requestOptions.value = {
+  if (latLng.value && resultsElem.value) {
+    resultsElem.value.requestOptions = {
       headers: {
         'x-user-location': latLng.value,
       },
@@ -68,7 +77,9 @@ onMounted(() => {
   }
 
   nextTick(() => {
-    pause.value = false;
+    if (resultsElem.value) {
+      resultsElem.value.pause = false;
+    }
   });
 });
 
@@ -79,63 +90,24 @@ eventBus.$on('sticky-nav-off', () => (stickyNav.value = false));
 
 searchForm.defaults(props.parameters).reset();
 
-const isNotEatery = (type: SearchableItem): boolean => {
-  return type !== 'Eatery' && type !== 'Nationwide Branch';
-};
-
-const formatDistance = (distance: number): string => {
-  const roundedDistance = distance.toFixed(2);
-  const label = pluralise('mile', distance);
-
-  return `${roundedDistance} ${label} away`;
-};
-
-const itemTypeClasses = (type: SearchableItem): string[] => {
-  const base = [
-    'rounded-lg',
-    'px-2',
-    'py-2',
-    'leading-none',
-    'text-xs',
-    'font-semibold',
-    'lg:text-base',
-    'lg:px-4',
-  ];
-
-  switch (type) {
-    case 'Blog':
-      base.push('bg-primary');
-      break;
-    case 'Recipe':
-      base.push('bg-primary-light');
-      break;
-    case 'Eatery':
-    case 'Nationwide Branch':
-      base.push('bg-secondary');
-      break;
-    case 'Shop Product':
-      base.push('bg-primary-other text-white');
-      break;
-  }
-
-  return base;
-};
+const shouldLoad = ref(true);
 
 const handleSearch = () => {
-  pause.value = true;
+  shouldLoad.value = true;
+
+  if (resultsElem.value) {
+    resultsElem.value.pause = true;
+  }
 
   submitSearch({
     onSuccess: () => {
-      pause.value = false;
-      reset();
-    },
-  });
-};
+      if (resultsElem.value) {
+        resultsElem.value.pause = false;
+        resultsElem.value.reset();
+      }
 
-const goToEaterySearch = () => {
-  router.post('/wheretoeat/search', {
-    term: props.parameters.q,
-    range: 5,
+      shouldLoad.value = false;
+    },
   });
 };
 
@@ -162,12 +134,11 @@ watch(
 watch(
   () => latLng.value,
   () => {
-    console.log('here');
-    if (!latLng.value) {
+    if (!latLng.value || !resultsElem.value) {
       return;
     }
 
-    requestOptions.value = {
+    resultsElem.value.requestOptions = {
       headers: {
         'x-search-location': latLng.value,
       },
@@ -250,123 +221,31 @@ watchDebounced(
       </Card>
     </div>
 
-    <Card
-      v-if="searchForm.processing"
-      class="w-full mt-4!"
-    >
-      <Loader
-        color="primary"
-        :display="true"
-        :absolute="false"
-        size="size-12"
+    <Deferred data="results">
+      <template #fallback>
+        <Card class="w-full mt-4!">
+          <Loader
+            color="primary"
+            :display="true"
+            :absolute="false"
+            size="size-12"
+          />
+        </Card>
+      </template>
+
+      <SearchResults
+        v-if="results"
+        ref="resultsElem"
+        :should-load="shouldLoad"
+        :results="results"
+        :landmark="landmark"
+        :has-eatery="hasEatery"
+        :location="location"
+        :term="parameters.q"
+        @mounted="shouldLoad = false"
       />
-    </Card>
+    </Deferred>
 
-    <Card
-      v-else-if="items.length === 0"
-      class="w-full mt-4!"
-    >
-      <div
-        class="py-8 px-4 text-center text-xl font-semibold text-primary-dark"
-      >
-        No results found!
-      </div>
-    </Card>
-
-    <div
-      v-else
-      class="group xmd:pt-2 xmd:-ml-3! flex flex-col space-y-2 min-h-screen"
-    >
-      <Card
-        v-if="hasEatery"
-        class="mx-4 rounded-xl border-2 border-primary"
-      >
-        <p class="prose lg:prose-xl">
-          If you're looking for places to eat in
-          <strong v-text="location" />, you can find more detailed results in
-          our
-          <a
-            class="inline-block font-semibold cursor-pointer"
-            @click.prevent="goToEaterySearch()"
-          >
-            Eating Out guide
-          </a>
-        </p>
-      </Card>
-
-      <Card
-        v-for="item in items"
-        :key="item.link"
-        class="transition-all transform scale-95 hover:scale-100 p-4 group/item group-hover:opacity-50 hover:opacity-100!"
-      >
-        <Link
-          :href="item.link"
-          class="flex flex-col space-y-4 sm:flex-row sm:space-x-4 sm:space-y-0"
-          prefetch
-        >
-          <div class="w-full sm:max-xl:w-1/4 xl:w-1/5">
-            <RecipeSquareImage
-              v-if="item.type === 'Recipe'"
-              :src="item.image"
-              :alt="item.title"
-            />
-            <img
-              v-else-if="isNotEatery(item.type)"
-              :src="item.image"
-              :alt="item.title"
-            />
-            <StaticMap
-              v-else
-              :lng="(<EaterySearchResult>item).image.lng"
-              :lat="(<EaterySearchResult>item).image.lat"
-              :can-expand="false"
-            />
-          </div>
-
-          <div class="flex flex-col space-y-4 sm:flex-1 sm:space-y-2">
-            <h2
-              class="text-primary-dark text-xl font-semibold group-hover/item:text-black transition lg:max-xl:text-2xl xl:text-3xl"
-              v-text="item.title"
-            />
-
-            <p
-              v-if="typeof item.description === 'string'"
-              class="prose max-w-none lg:prose-xl flex-1"
-              v-text="item.description"
-            />
-
-            <div
-              v-for="(restaurant, index) in item.description"
-              v-else
-              :key="index"
-            >
-              <p
-                class="prose-xl max-w-none lg:prose-2xl flex-1 font-semibold"
-                v-text="restaurant.title"
-              />
-              <p
-                class="prose max-w-none lg:prose-xl flex-1"
-                v-text="restaurant.info"
-              />
-            </div>
-
-            <div class="flex justify-between items-end mt-auto">
-              <div
-                :class="itemTypeClasses(item.type)"
-                v-text="item.type"
-              />
-              <div
-                v-if="(<EaterySearchResult>item)?.distance"
-                class="text-sm text-grey-off-dark group-hover/item:text-grey-dark lg:text-base"
-                v-text="
-                  formatDistance(<number>(<EaterySearchResult>item).distance)
-                "
-              />
-            </div>
-          </div>
-        </Link>
-      </Card>
-    </div>
     <div ref="landmark" />
   </div>
 </template>
